+++ R/widgets.R
@@ -1,15 +1,33 @@
 #' HTML Widgets for graphical comparison
 #'
+#' These widgets can be used at the console and embedded in a R
+#' Markdown document or Shiny application.
+#'
+#' The regular versions take plots or functions as \code{before} and
+#' \code{after} arguments (see \code{\link{expect_doppelganger}()}
+#' for details). The versions suffixed with underscores take HTML image
+#' sources. These can be paths to SVG files or inlined SVG
+#' images. Currently, \code{widget_diff_()} is compatible only with
+#' inlined images.
+#'
 #' @inheritParams htmlwidgets::createWidget
-#' @param before Path to the picture that is taken as reference.
-#' @param after Path to the picture against which the reference is
-#'   compared.
+#' @param before The picture that is taken as reference.
+#' @param after The picture against which the reference is compared.
 #' @name htmlwidgets
+#' @examples
+#' p1 <- function() hist(mtcars$disp)
+#' p2 <- function() hist(mtcars$drat)
+#'
+#' # You can also call these functions in a R Markdown document or
+#' # in a Shiny application:
+#' widget_toggle(p1, p2)
+#' widget_slide(p1, p2)
+#' widget_diff(p1, p2)
 NULL
 
 #' @rdname htmlwidgets
 #' @export
-widget_toggle <- function(before, after, width = NULL, height = NULL) {
+widget_toggle_ <- function(before, after, width = NULL, height = NULL) {
   sources <- list(files = list(before = before, after = after))
 
   htmlwidgets::createWidget("vdiffr-toggle",
@@ -22,7 +40,7 @@ widget_toggle <- function(before, after, width = NULL, height = NULL) {
 
 #' @rdname htmlwidgets
 #' @export
-widget_slide <- function(before, after, width = NULL, height = NULL) {
+widget_slide_ <- function(before, after, width = NULL, height = NULL) {
   # Drawing a SVG into a canvas requires that the svg node has 'width'
   # and 'height' attributes set. Otherwise the result is oddly cropped.
   sources <- list(before = before, after = after)
@@ -38,7 +56,7 @@ widget_slide <- function(before, after, width = NULL, height = NULL) {
 
 #' @rdname htmlwidgets
 #' @export
-widget_diff <- function(before, after, width = NULL, height = NULL) {
+widget_diff_ <- function(before, after, width = NULL, height = NULL) {
   sources <- list(before = before, after = after)
   sources <- list(sources = map(sources, svg_add_dims))
 
@@ -48,4 +66,40 @@ widget_diff <- function(before, after, width = NULL, height = NULL) {
     height = height,
     package = "vdiffr"
   )
+}
+
+#' @rdname htmlwidgets
+#' @export
+widget_toggle <- function(before, after, width = NULL, height = NULL) {
+  files <- widget_svgs(before, after)
+  widget_toggle_(files$before, files$after, width, height)
+}
+
+#' @rdname htmlwidgets
+#' @export
+widget_slide <- function(before, after, width = NULL, height = NULL) {
+  files <- widget_svgs(before, after)
+  widget_slide_(files$before, files$after, width, height)
+}
+
+#' @rdname htmlwidgets
+#' @export
+widget_diff <- function(before, after, width = NULL, height = NULL) {
+  files <- widget_svgs(before, after)
+  widget_diff_(files$before, files$after, width, height)
+}
+
+widget_svgs <- function(before, after) {
+  out <- suppressMessages(list(
+    before = svglite::stringSVG(print_plot(before, "")),
+    after = svglite::stringSVG(print_plot(after, ""))
+  ))
+
+  # widget_diff() does not work if SVG doesn't finish with newline
+  out <- map(out, paste0, "\n")
+
+  # Inline SVGs so the widget can be easily embedded anywhere
+  out <- map(out, as_inline_svg)
+
+  out
 }